gh-426 connect List of Things resolver in GraphQL with db connnector
As of now the db connector is still returning fake data and a lot of
over-simplifications are being made right now. For example, we're
currently pretending that the result of a Thing query is a
map[string]string and the resolver for the individual fields isn't
capable of handling any other type as of now.

Additionaly, there is now a GraphqlListThings method on the dbConnector
interface. I'm proposing (more later on the PR for this branch) that we
will have individual methods for each use case very similar to the REST
API. A basis for that discussion has already been started on #480