gh-426 query janus graph when resolving ListThings in dbConnector
The previous commit simply returned dummy data, now we actually have
data returned form the graph db. Most of the over-simplifications from
the previous commit still apply in similiar fashion.

Previously we assumed, every primitive type that has to be resolved is a
`string` as part of a `map[string]string`. This has now changed to a
`map[string]interface{}`. However, I don't think we have to keep this
oversimplification, as we actually have more precise types for fields in
the models.

Nevertheless, ignoring the actual type was a quick way to get a first
MVP (query the name of all cities from the DB through graphQL). And it's
serving it's purpose.