gh-494 decouple connector from graphql api
This commit changes the proposal from the previous commits to make sure
the dbConnector has no more knowledge of the graphQL API existing.

To achieve this we basically just reuse existing non-Graph methods on
the dbConnector interface. The resolver functions (graphql logic) are
now in the graphql package. If we were to add a third API (other than
REST and GraphQL) no dbConnector would have to be touched.

Over-simplifications:
- The existing ListThings methods doesn't filter for the atClass, so
currently we're not really filtering for a class (easy to add though).