gh-602 prepare network pluggin, inject network to graphql build
We are now providing graphqlapi.Build with not just the db (acting as a
DatabaseResolverProvider), but also with the Network (acting as a
NetworkResolverProvider). The combination of the two is now the upated
interface ResolverProvider..

To resolve some import cycles we have moved much closer to a pluggable
entwork structure, but it is not fully pluggable yet, as we have (or
will have) a separate issue for this.