gh-602 forward auth headers to remote instance on Network.Get
We're using an http.RoundTripper, which basically acts like a
middleware, it is called before the request is sent off. This allows us
to inject the X-API-{TOKEN,KEY} headers. We are getting those in turn
from the grahpql context.

The current setup simply calls http.DefaultTransport after the custom
RoundTripper is done. This will work as long as we only have one custom
RoundTripper. Having said that I don't currently see a reason why we
would have more than one in this particular place...