gh-602 proxy filters to remote instance on Network.Get
Getting the parsed filters back to their source string was not as easy
as anticipated. The graphql-go package parses them into
map[string]interface{} which is fine to read them.

However, this presents us with two problems:

1.) When we want to "marshal" them back into the graphql format, the
format is like JSON, but is not actually JSON. The keys can't be quoted.

2.) String values are quoted, as expected, but not every value is a
string value from the perspective of the graphql schema. Operators for
example are an enum, so they also shouldn't be quoted.

This problem is currently solved with two regular expressions, to strip
the quotes off of all keys and additionally to strip the quotes off of
the values for an operator key. Not the best-scaling approach, but the
where-Filter is luckily a relatively small and predictable set of
fields. We can always look for a more elaborate solution if we actually
run into issues because of this.