gh-602: in acceptance tests always have weaviateB and C in network
This is a rather interesting one. The acceptance tests added as part of
issue #505 re-enables the graphql test that compare an introspection of
the prototype and the actual implementation.

This also means that the exact same network instances have to be
present. Prior to this feature this was a fixed list, but now it's a
dynamic list of which instances are currently registered in the network.

As a fix, we're now adding the expected instances through our genesis
fake. (Good thing it's a fake and easy to control ;) )