gh-609: make current schema accessible in network ping cycle
To reduce coupling a RegisterHandler was added to the network
interface, every network must implement this handler and use the getter
function that is being passed to retrieve the current schema.

This happens asynchronously (and repeatedly) which is why the "register
getter" way was chosen - so that the getter can be called later.

The signature of the passed getter is a small interface ("Schema
Getter") which tells the underlying type to provide a Schema() method
which in turn provides the current schema.

Most likely when a schema should be retrieved locking is involved,
however the network wants to stay agnostic of the internals of how a
schema provider deals with locking. In turn any type that wants to
implement the SchemaGetter, should take care of locking as part of the
Schema() method it provides.