gh-609: extract Peer(s) into own packages
This will make it much easier to transport the schema (which is part of
the peer) into the graphql api package. As of now that package is
agnostic of peers, only cares about a []string for their names. But with
the increasing complexity of peers having a hash and a schema this
attempt to decouple doens't benefit us as much as simply having an
isolated structure to describe peers.