gh-609: in graphql api use peers schema instead of local schema
Previously we were taking a shortcut where we'd assume that every
peer has the same schema as the local version.

The commits leading up to this one made sure that a cached version of
another peer's schema is available to the graphql API. This commit now
also builds the graphql fields based on that particular schema.