gh-609: add acceptance test for remote peer w/ differnt schema
We already had an acceptance test for a Network.Get from a remote peer.
However, since previously all peers had to have the same schema, the
tested schema also was the same as on the local instance.

The new test explicitly uses a different schema now to highlight that
they can be different. In order to make the tests pass the remote peer
fake now also offers that particular schema on the schema enpdoint.