gh-642: skip test which requires host access on travis
We would have to improve our test setup a bit before this test also
works on CI.

Namely, we'd have to make sure that the tests themselves also run inside
docker so that they're in the same network as the genesis.

In the meantime, we can at least run them locally.