gh-627: udpate user-facing p2p documentation
It was outdated in two ways:

- It showed an old network architecture model, where a genesis did
exist, but wasn't a standalone product. But rather a peer would turn
into a genesis and then turn into a regular peer. This has been
completely rewritten now to indicate that the gensis is what makes the
network a hybrid-p2p network.

- It showed outdated configuration options for an unknown field `p2p`
in the local configuration. This has been corrected to use the correct
field `network` now.