gh-630: sketch out unit test structure for graphql builder
Right now we have an issue where duplicate schemas can lead to duplicate
object, as we found out after moving the where filter.

It's very hard right now to get feedback for that, as you basically have
to run the entire acceptance suite, with actual peers. A unit test
should be much easier to provide feedback on that. However, we first
need the changes from PR #649 becasue they substantially change the
method signarture with regards to what peers are. Previously there were
just a []string and afterwards they'll be a complex object with
individual schemas attached.