gh-658 support 'cardinality:many' on import of network ref
For some reason the previous code used a different method to add many
refs to single refs. The single-ref code had extensive validation (for
local refs) that was skipped for many-refs.

The new code simply loops over the many refs and for each runs the same
code that we have for a single ref. This means the switch between
network and local as well as the additional validation for local refs is
now exactly the same.

The importer script previously wasn't capable of adding refs with
cardinality many. This has been fixed, but only for things, as part of
this commit.

We need to identify whether the current importer should become a
strategic tool (possibly as part of the weaviate-cli) or if it's just a
helper for our local dev setup and tests. If it's the latter it's good
the way it is, but for the former we would have to improve upon it quite
a bit.