gh-658: validate remote kind exists on import
To make this work with the dev-setup we are again relying on the
weaviate_b_fake which is simply the graphql prototype. In order to pass
validation the prototype will respond for exactly one "Country" with a
specific ID.

Additionally, we are starting to uphold the convention that a local
reference must contain 'localhost' (without ports) as the host,
regardless of what network address it's actually running or reachable
on. The logic is now simply "localhost" == local ref. All other hosts,
even 127.0.0.1 represent a network ref, referenced by peer name (!), not
actual host or FQDN.