gh-658: implicitly update action schema on add action
We previously only supported this when an action was created without a
network ref, but then a network reference was added through a patch. Now
it is also supported if the network reference is present right from the
get go. (The same is not yet the case for things, but will be added
soon.)