gh-658: add missing unit tests for referenceSchemaUpdater
Since the udpater lives in a mostly untested context, I initially
assumed that it would be too costly to build it in a test driven
fashion. However, as it turns out, it is very well decoupled and there
was no more reason not to have a unit test for it.