gh-659: build graphQL from peer schemas for network classes
Since this functionality is very decoupled from the rest, we chose to
put it into a separate package. It does however share some logic with
other parts of the graphql packages, namely the building of primitive
props. As a quick solution this code was simply copied over from the
local_get package. Later on, we should see if this same switch-case
statement repeats over and over and extract it into a separate package
that can then be used throughout the graphql tree.