gh-659: include network refs in graphql
As outlined in #413 we aren't capable of resolving the network refs yet,
that will be tackled in #660.

Since we don't resolve yet, we also can't test this feature as a journey
test yet. However, since we have too many of those anyway, it is now
tested with a unit test (technically rather a component test) at the
local_get level. There was already an existing component test for
local.Get without network references. Having the combination of both was
a good helper to test that everything still works, both when network
peers are present and when they are not present.

Being able to run those test without any dependencies in under a  second
is a massive feedback cycle time reduction compared to the acceptance
tests which take about 10 minutes.