gh-630: fix resolvers for local get where filters after move
The filter are now usable again, after being moved to the KindClass.
Previously they were at the level of Local.Get which didn't help
anything, but created plenty of issues that can be read in the issue
description (#630).

This also happens to fix the issue about not having enough arguments.
Therefore this commit fixes #629.