gh-660 don't return thunk for connecter.LocalGetClass
This seemingly simple change has quite some major implications. We were
returning a thunk previously to avoid our results bubbling up to higher
resolvers.

This behavior was a bug in go-graphql that has been fixed in v0.7.7:
https://github.com/graphql-go/graphql/issues/340 We are currently on
v0.7.5.

This commit removes the thunk and simply returns the results directly,
in order to make this work we had to introduce a new resolver at the
level of the reference field. Additionally we needed to add a
ResolveType function at the level of the classUnion (that is used to
for the reference field).

This has a nice side-effect however: We can also use the resovler to
determine whether it's a network ref or a local ref and in the case of
the former send out a query to the remote peer. (This is not in this
commit yet, this is the goal of this issue)