gh-660 only query peer if user wants classes from this peer
Note: (copy paste form a code comment)

---
We cannot do an exact check of whether the user specified that
they wanted this class as part of their SelectProperties because
we would only find out the exact className after resolving this.
However, resolving a network ref is a) not the concern of this
particular database connector and b) exactly what we want to
avoid. The point of checking whether this class is desired is to
prevent unnecassary network requests if the user doesn't care
about this reference.

So instead, we can at least check if the user asked about any
classes from this particular peer. If not we can safely determine
that this particular network ref is not desired.
---

Additionally this commit includes a test file that I previously forgot
to commit and some refactoring on the janus connector
LocalGetClassResolve which was another massive method that has now been
split up a bit to make it more readable and extendable.