gh-680: spike out happy path for a single primitive filter
In order to do so, the previous filter implementation has been removed
entirely. To my surprise not a single high-level test fails because of
this. :(

What does this hard-coded POC do?
- always assume there is just a single filter condition
- always assume the filter condition is "lessthan"
- always assume the filter prop and its value is a number
- completely ignore reference props

If all these conditions are met, we actually have a working
implementation. These means the spike was a success and we now have a
relatively clear path ahead to build working filters for
- all primitive props
- combinations of primitive prop filters

Since this is a relatively well-isolated concern, I will build the new
logic in a sub-package and make sure the new code is test-driven, thus
hopefully also improving the quality of the current implemenation.