gh-653: add type inspector to retrieve type of ref props
The type inspectors results are merged with the janus processors
results, so that we can neatlessly query regular getmeta info and type
info even though the come from completely different source (types com
from the schema provider (currently saved in a file), whereas GetMeta
info comes from aggregation functions in the janus db.

This type inspector is currently only used on Reference Props, however
the plan is to make it usable on every prop, so that it can inject the
'type' field when the user requests it.