gh-653: enable support for where filters
Thanks to the way we built the filter query builder in a very isolated
and pluggable fashion in #680, we could simply plug them in right here.

Interestingly the first (explorative) tryout of the filters highlighted
a small issue that I previously hadn't thought about. To calculate the
percentage of bool values we rely on having a count for true and false.
However, while using the filteres I narrowed down the result sets of
Cities where all results would have 'isCapital == true'. This led to an
error in the meta aggregator:

  got a successful response from janus, but could not merge the results:
  could not postprocess 'map[string]interface
  {}{"boolGroupCount":map[string]interface {}{"true":1}}': boolProps did
  not contain any info about no of props which are 'false', got
  'map[string]interface {}{"true":1}

However, filters or not, it is of course completely viable to have
result sets where all boolean props are true, or all boolean props are
false.

So, we have to introduce a fix for this in an upcoming commit.