gh-653: add simple gremlin playground
modify the file query.groovy and run the submit.go script. It will
pretty-print the JSON results from Janus. This is very helpful in trying
to debug Janus queries.

It makes a lot of sense to combine this with a file watcher, for example
'reflex'. Then you can run your watcher like so, if you are located in
the ./tools/dev/gremlin-playground/ folder:

$ reflex -- go run submit.go