gh-684: overwrite janus config to enable batch loading
This is so that we are independent from waiting for a potential new
release of the janusgraph-docker image. (Full disclosure: We own the
repo that builds that image as well, however, I'm a bit conflicted with
the versioning strategy there at the moment, since this is a potentially
breaking change and right now the image tag simply corresponds to the
janus version. I.e. we'd have no reason for a bump on a config change.)