gh-653: enable <refProp>.count
I must have accidentally ticked that off on my todo list, because it was
ticked off but when recording a demo video I realized that the feature
was never built. It is now.

Note that there's currently a bit of confusion about capitalization of
refProps, such as "InCountry". On user-facing parts we have them
uppercased, but on schema imports (and therefore in the database) they
are lowercased like "inCountry". I don't really see a reason for that
and it currently leads to some ugly transformations, such as
string.Title() being used, when displaying the prop and untitle(), i.e. a
custom untitling function, being used when taking in user input.

I want to discuss and potentially fix this confusion as part of #677.