gh-646: remove key auth from requests
unfortunately this meant that the auto-generated code from swagger
changed the method signatures slightly. As we don't have an auth scheme,
there's also no point for go-swagger to pass around the principal all
the time. Once we re-add auth, we'll probably have to change the method
signature again, however, it's relatively well automatable.

We have not removed the key-specific API endpoints yet.