gh-644: (refactor) clean up mixed concerns in network package
The network packages has sub-packages for each new branch in the tree,
such as Get, GetMeta, aggregate, etc. However, while tyring to write a
component test for GetMeta, we learned that in fact some of the GetMeta
logic (and that of other subpackages as well) is not in fact in the
seemingly isolated packages, but rather mixed up in the network
(root) package.

This refactor commit moves those separate concerns into their individual
packages. To make sure we don't break anything a few component tests
were added, that assert the successful creation of the graphql schema.
Since this part was previously untested on anything other than a e2e
level, we discovered lots of edge cases that would lead to errors, e.g.:

- A peer with an empty schema could block the creation
- A peer with a class without properties would lead to errors in the
Aggregate package, which blocked the creation
- etc...

All errors we discovered through the new component test were fixed.