gh-644: fix "not supported" resolvers in network GetMeta
This is verified with a component test. This means we have proxying
ability from now on. However, there seems to still be an issue with
number fields, as numbers are typed as json.Number, graphQL doens't seem
to be able to handle those with the default resolvers for number types.