gh-644: resolve json.Number correctly in graphQL api
Since the results are proxied from another peer and we use the
go-graphql client, the numberes are typed as json.Number. This commit
makes sure that every conceivable (unless I missed one in my tests)
GetMeta field is able to deal with the json.Number type.

The nice side-effect is that we have a component test now that tests a
lot of edge cases, i.e. almost all possible properties. This can
potentially also prevent us from unrelated regressions.