gh-693: build query to aggregate numerical props with count
so far, we can only aggregate the count, nothing else yet. Additionally,
we are currently pretending that grouping is always done by primitive
props. Additionally no post-processing is happening yet, however
post-processing will be necessary as we are currently returning
aggregation keys such as "prop_18__count", this is because they have to
be unique in our gremlin query across all props, but will repeat in the
graphql api.

These limitations will be overcome with future commits.

Additionally, please note that we already have edge case handling for
queries with just a single property and queries with just a single
aggregator. The reaon for why these are edge cases are extensively
documented in database/connectors/janusgraph/aggregate/query_builder.go