gh-693: connect graphql API and janusgraph for Local Aggregate
Since we don't have a post-processor yet, we can't actually serve
queries yet, thus we simply print the results to stdout at the moment.
Nevertheless making this connection is very helpful, as we can now
assert (almost) end-to-end that the queries we have only unit-tested so
far actually do what they are supposed to.