gh-685: extensively document behavior of distributed RWMutex
With a local Mutex it's very important to pass it by reference, so we
always operate on exactly the same lock. For this distributed lock using
etcd it's the other way round. The lifespan of a Mutex type should only
be from Lock() to Unlock() or from RLock() to RUnlock().

This commit adds extensive documentation including examples of why not
paying attention to this is problematic.