gh-412: run unit tests first in test suite
This is to give early feedback if a unit test is broken, since we can
get this feedback before having to spin up the whole docker setup.

Currently the unit tests are run twice, i.e. when the acceptance tests
run they run again, because here we also create an extensive
coverprofile. However I'm not sure if this is even used, as code_climate
seems to be turned off. Nevertheless this was useful tool and I'd be
happy to turn this on again.