gh-685: always take in context from the outside
The etcdv3 client requires the ctx as one of its parameters, this is a
good pattern that we should also aim to keep up. Ideally, the restapi as
constructed by go-swagger is the creator of a context and then
everything that happens as a result of an incoming rest request should
pass along the context. We are now passsing it along for everything
schema related already, but it would be nice if everything (database
connector read/write operations, network operations, etc) in the future.