gh-685: reenable schema manager tests
Previously they needed a temporary file directory to write to the disk.
Since changing from the local implementation to the etcd implemenation
we previously weren't able to run this test anymore. As part of the POC
this was acceptable, but now when moving to production-quality code, we
need it again of course.

I saw two options:
a) Turn this into an integration test
b) Fake etcd.

Since we were using only two methods on the etcd client (Get() and
Put()) faking was the much cheaper option. The fake is a very simple
map[string]string as a KeyValue store which gives us everything we need
to make the tests pass again :)