gh-645: build simple schema search onto contextionary
Previously the "search" interface on the contextionary would simply be
to run a kNN search on it. however, this would mean that you would
handle raw words and would not have any indication of whetehr they are
class names or properties or just words in the contextionary.

The schema search adds that abstraction and makes it much easier to use
from the rest of the app, where no knowledge of the internals of the
contextionary should be required.