gh-645: error in graphql resolver if c11y search returned no results
The two special cases that we can run into are if either there is no
class found or we have classes, but don't have any properties to search
on. This would lead to errors down the line in the connector. However,
instead of having to handle this special case in every single connector
we're handling it in the earliest possible layer now, i.e. in the
graphQL resolver.

This means that the database connector now has the guarantuee that there
will always be at least one class and one property to search for. This
gives now guarantuee that they "match" though. I.e. it could still be
possible that we get classA and property classB.propertyFoo back.