gh-645: prevent invalid queries on invalid class/prop combination
The graphqlapi package is already catching cases where no classnames or
no property names are valid at all. However, there is also another case
that can only be caught in the connector (since this is the first time
we validate against the schema). If the classnames themselves are valid
and the properties as well, but they aren't combinable or don't match
the specified filter (e.g. wrong type), then we can't build a query.
This woudl previously lead to an invalid gremlin query, now we are
instead catching the errror and trying to return a helpful message to
the user.