gh-715: replace remaining comoposite indices with mixed index
For this we need to make sure that we index string props either as
STRING or TEXTSTRING, meanin that we can also use the mixed index for
"exact-match" style searches. Previously whenever the gremlin query was
`eq(string)` it would fall back to a composite index even if other
paramaters where included that would have been better answered with the
mixed index.

We now only have a single composite index left: for uuid - guaranteeing
to make uuid unique, somethign that a mixed index cannot do.

This should improve performance at scale considerably.

Note that indexing with the external index is not always instant, so we
had to adjust some tests which were implicitly assuming an atomic
create-and-index process. They are now waiting for a created thing or
action to actually become available on the index before performing
further operations on them.