gh-713: POC (!) try different batch import strategy
Previously we would parallelize batch requests by spawning different
goroutines per item and send one request to janus. At scale this leads
to a massive HTTP overhead and a lot of network traffic.

This POC tries to send a single request per batch to the connector which
includes all vertices to be added.

WARNING: This is very much POC code that is NOT production ready. If it
proves to work well, it will get imporoved.

Things that are currently keeping this from beign production ready are:
- only doing this for things, actions is still the old way
- we don't return meaningful data
- validation results are completely ignored