gh-647: add acceptance test, change priority when creating predicates
See this inline comment, for we have changed the order/priority of
assembling the predicates:

Note that we are first iterating over the term, then over the prop. This
is because the terms are already an ordered list where the match with
the highest certainty is the first item. Janus also imposes a limit, so
we need to abort if we get too many predicates. Assuming we have 20
props and 20 matches, by iterating over the matches first, this way,
we'll be able to cover about the 6 best matches on all props. If we
split the order of iteration, we would cover all matches, but completely
ignore 14 out of 20 props.