gh-831 fix issues with deleting properties
Previously we would first try to remove the property from our own cached
(and stored in etcd) schema representation, then remove it from the db.
The latter would fail because it tries to query to former about the
property.

This would leave us in an inconsistent state, creating all kinds of
other fallout.

While fixing this, we discovered that plenty of other endpoints in both
REST and GQL would fail after a property was deleted. This should also
be addresesd with this now.

closes #831