gh-868: fix broken error handling for patch semantics
See
https://github.com/semi-technologies/weaviate/issues/868#issuecomment-491761905
for the actual cause of the issue.

Nevertheless this cause was hidden due to the fact that the error
handling for the update kinds UC was broken. This commit fixes the
broken error handling and adds helpful debug logs (only logged when
debug mode is running).

Additionally this commit adds a test suite replicating the steps from
the original issue to outline a working method which can be used as a
reference. Furthermore it guarantuees regressions, of course.