gh-873: don't run unit tests twice
previously unit tests where run once in the beginning and then as part
of the accpetance tests all packages were ran again. Because of the
-race flag, even an empty package would take considerable time. This
makes sure the unit tests are only run as part of the unit test section.
The acceptance test section then only runs what is in the approproate
folder.

Additionally, this starts removing obsolete layers from the Dockerfile,
however more investigation is required to make it really efficient.