gh-730: merge dev and prod build process
Previously the c11y inside the weaviate container was a deciding factor
between using a dev or production image. Now we can finally get rid of
this big anti-pattern. Since configuring the contextionary is simply
pointing it to the right URL now, there is no more need to have separate
proceses. Instead we can just use one image for weviate and change the
contextionary image as we like

This simplifies the setup quite a bit. (And should lead to a massive
reduction in build times)