gh-891: use has(classId,x) intead of hasLabel(x) in Janusgraph
It seems - through trial and error - that we could identify that the
has("classId", "foo") syntax works effeciently with the ES index at
medium to large scale. Whereas the .hasLabel("foo") syntax lead to very
inefficient queries that didn't make use of ES's filtering abilities,
but instead iteration over thousands of entries to filter the results in
Janusgraph.

fixes #891