gh-1069 use smarter request cacher
Instead of recursively contacting the backend, we are checking for the
first level of references and then fetching all desired refs in a single
request. For my test case this already got the response time down from
350ms to 60ms for a single level request.

Note that resolving multiple levels is not possible yet. And that the
code requires quite some cleanup.