gh-1069 on deduping ignore complete jobs
A job might already be marked as complete, but when checking references
we might come across the same job again, then missing it in our nested
cache. As a result, we're now ignoring completed jobs. This leads to an
additional lookup (but not request) in some cases, but prevents an
impossible to resolve situation in others.

Note: The lookup is not yet ideal, as we seem to overwrite elements that
appear twice in the list with different select properites. (e.g. at the
root level with a lot of nested props and a second time due to a
circular reference.) We can address this, by not selecting solely by ID,
but by a combination of ID and select props.