gh-1125 prepare contextual prepare run
A prepare run is a run done before the individual per-item
classifications to provide an additional context. As part of this
prepare run we plan to fetch information that stays valid for the entire
run, such as target vectors. Additionally this is the place to calculate
tf-idf scores over all documents.