gh-1125 temporarily extend c11y client with MultiVectorForWord
This should be an actual batch method in the c11y. This is just a
temporary shortcut to move a bit faster. However, this is very
problematic as we still have a lot of request overhead and have no clean
way to determine word-not-found errors at the moment.

This will be addressed before completing #1125