gh-1218 support aggregation for kNN classifications
Note: There is currently a mix-up of business and persistence logic as
the methods demanded by the classification UCs require there to be too
much business logic in the persistence package. However, they have been
built in a very general way, so that the logic can be moved into the
respective UCs in the future.

However, due to the nature of how much logic is touched by the
classifications, the integration tests still add a lot of value. So once
we move the business logic out, we should keep the tests and simply
replace the interface with the respective UCs so we can still test the
persistence logic related to it. This will effectively make the tests a
component test.