gh-1219 prevent commitlogger/condensor from overflowign uint16
it seems in very large use cases we have a lot of links, more than can
fit in a uint16. If it overflows, the commit log gets corrputed.

This probably has a root cause, I don't think we should have this many
connections on any node. But for now that's just a workaround to prevent
us from writing unreadable commit logs