gh-1249 use geo index in searches on geo props
Note 1: At this point we don't yet have anything that writes into this
index

Note 2: Because of a global search-n-replace this commit also contains a
gofumports of the entire code base. This is something that was overdue,
but it's a bit unfortunate that this made it into an unrelated commit
now.