gh-1272 [ci skip] fix merge issue where vector was ignored
The merge implementation relied on the fact that - in the old
implemtation - an explicit vector update was the only thing that could
lead to a new doc id. Now that every update leads to a new doc id, the
old implementation would sometimes write into the vector index with a
nil-vector

Note that we currently still have the same issue on the batch-reference
which also relied on the fact that a ref update would never alter the
vector position. It is marked with a TODO and needs to be fixed before
the completion of this issue.