gh-1272 fix additional index races leading to flakyness
In the reference batcher we previously attempted concurrent reads/writes
which aren't a problem per se. However, we could get into a situation
where we would mark a doc id as deleted before it was ever imported.
That would be problematic as doc ids on additional indices (such as
HNSW) are themselves immutable. So an import to a deleted doc id would
lead to issues, as we couldn't distuingish this from an attempt to reuse
a doc id.

As a side-effect the fix for this also made the ref-batcher slightly
more efficient, as we remove unnecessary updates on the "same" object
which was touched multiple times in one batch and therefore received
multiple increases on the doc id