gh-1342 add nearText & nearVector to GraphQL
Note: The uc/explorer is not yet capable of handling the new types at
this point and must be adapted in a future commit.

It seems the Explore API is missing a journey test. This commit should
not be green as Explore { } is currently not working, yet somehow it
is...