gh-1334 keep doc id on batch ref updates
They are a special case because we can be sure that no inverted index
clean ups are required (as we only add to it) and no vector indices
require udpating as a ref does not alter the vector position.

This allows for a considerably faster import. The news-pub dataset which
I used for testing while development was imported in around 1:45min
instead of the previous 2:30min.