gh-1559 [skip ci] use random vectors for large test cases
It seemst he previous implementation (incorrectly) ignored with the
exact same distance. In reality this is rare or even impossible, so we
never ran into this, but in tests fixed vectors where often hard-coded.
This commit makes sure they are random as to prevent having to perform
massive exhaustive searches on them