gh-1559 set graph integrity test to require at least 1 connection
As outlined in the the test's comment already it is very questionable
how much value this test adds with the use of Heuristic2 in HNSW. The
heurisitic skips "useless" connections to make sure only the key
connections are kept. In extreme cases this can mean that only one
connection is kept. At least the test will still error if nodes are
entirely unconnected which would indicate an issue.