gh-1569 [skip ci] WIP fix disk tree, start writing into inverted
The diskindex binary tree was extremely unbalanced leading to insane
memory requirements. This commit fixes it and makes sure it is built up
perfectly balanced. This also helps us get rid of the rather hacky
shuffling beforehand.

This commit also starts writing into the inverted index, but currently
no reading from it is done.

Note that this commit also switches various commit loggers (in both HNSW
and in the lsmkv) to be buffered io writers as the many parallel.Write()
syscalls became very problematic. However, this means that we need to
introduce forced flushes at various points, such as after completing a
batch, etc. This has not been completed yet.