gh-1599 instead of using single shard use first real shard
This does not yet give multi-sharding support, but it no longer uses a
fake shard, but a real multi-sharding config that has only a single
shard. Will error if more than one configured.

Note: At this stage this change would be incompatible with previous
releases, as it does not yet convert old shards to the new conventions.