gh-1652 flesh out a disk based hnsw condensor
This is a rough draft, I think the approach should work, but it's not
complete. I basically picked this up between two larger tasks and plan
on continuing it next time, I'm between tasks.

It uses an index-based approach and an mmap where the index is kept in
mem and the payload is mmaped. A rough calculation showed that with a
30M hnsw graph, the index would only grow to 600MB, whereas with the old
memory-based approach the mem requirements would considerably higher:

For example, with M of 64 and even ignoring levels higher than 0, at
30M objects the requirements would have been 30GB already, so the new
solution is about 99% more memory efficient