gh-1658 forward put object to remote shard
Note: We had to move storobj into entities - which makes sense given
its global application-wide role. However it depended on a UC struct
(traverser.AdditionalProperties) which in turn had to be elevated to an
entity. This also makes sense given that additional properties are a
core weaviate feature and are not specifying to traversing. They can for
example also be used with the REST API which does not use the traverser
at the moment.