gh-1658 resolve crossreferences (multi-get)
This required changing the payload structure of some of the existing
client/api methods where we incorrectly used JSON which would not be
unmarshalled into the correct types when reading. This has been changed
to use binary methods from storobj package, which means the type parsing
is consistent with what we get from disk.