gh-1740 make reading vectors much more efficient
Previously we were reading (and unmarshalling!) the entire object from
disk whereas now we are only looking at the bytes relevant for the
vector. This lead to a massive speedup for pre-filling.

This commit also fixes the faulty lock that would render the entire
index unusable during the pre-filling process. It now only locks for one
iteration.