gh-1872 clean up disk segment list properly on shutdown
This prevents a nil-pointer panic that showed as part of a flaky test.
The actual cause is described in:
https://github.com/semi-technologies/weaviate/issues/1872#issuecomment-1076357479

To assure the test is fixed you can run run

```go
go test -race -tags integrationTest -count 10 -run TestMemtableThreshold_Replace ./adapters/repos/db/lsmkv
```

which reliably failed before, but no longer does so with the fix.

fixes #1872