gh-1897 allocate new connList if more than allowed connections
How is it possible that we could ever have more connections than the
allowed maximum? It is not anymore, but there was a bug that allowed
this to happen in versions prior to v1.12.0:
https://github.com/semi-technologies/weaviate/issues/1868

As a result the length of this slice is entirely unpredictable and we
can no longer retrieve it from the pool. Instead we need to fallback
to allocating a new slice.

This was discovered as part of
https://github.com/semi-technologies/weaviate/issues/1897