gh-1908 reuse parsed collection node buffers (150GB->126GB)
This commit allows parsing a collection node into a reusable buffer.
Additonally it supports a reusable segment cursor for collections that
makes use of this functionality. As a final step this new cursos is used
as part of the map compactor.

This commit is the final step on this allocation reduction journey. Out
of the remaining 126GB of allocations maybe only 2-3 GB are left from
the map compactor. However, these also include things that stay on the
heap such as bloom filters. I would recommend to focus elsewhere
(memtable, flush cycle) etc for further improvements.