parse "replace" wals into cache first on recovery
This is because the current memtable implementation does not deal well
with extremely long commit logs with a lot of duplicate entries. The
cache is a hashmap that can be used to reduce the duplicates. By only
taking into account the unique values, parsing WALs with a lot of
redundant info is much faster.