Improve implementation of the visited list (fixes, new memory allocation strategy) (#2023)
TD-193: Improve memory allocations using a new strategy

Fix memory leaks

Improve implementation of the list pool and RC

This PR makes the visited list faster not necessarily reducing the amount of used memory.

This is because the same approach (set as []int8 ) is used as before but the implementation has been changed

The main advantage of the new implementation are:

The list is compact because it can be hold as a value (no pointer needed).
It reduces the number of times the list needs to be resized because it uses variable growth. This is the major difference to the current implementation which use fix growth (1000 bytes). This will come in handy when inserting huge number of objects.
Pool.Borrow() uses LIFO approach to manage lists and to prevent from memory leaks