delete new files when dropping a segment, fix remaining tests
The drop command is also called as part of compactions, so we can now
garuantee that no old state is left after compacting two segments.
However, currently we are also relying on the first segment init after
the compaction to build all files again. This works, but can probably be
optimized further:

My current assumption is that when two segments are merged, the cnas of
the existing segments can simply be added. Right now we would
re-calculate them at segment init which can be very expensive on large
segments.