Not really a solution for async
In the future we will have async full support. This being said, the commit introduces a callback to inform that the compression has finished. There exists way better ways to do this but since we plan a full support for async, I would rather wait for it instead of making a full solution here.