add more compaction tests
This bug surfaced a gap in our compaction tests. While we already had a
fairly eloborate test setup with various operations happening during
compactions, we only verified the correctness of the result using a
cursor. Bug #3592 was specific to accessing keys directly. Cursors would
have still worked fine, therefore the tests didn't catch this bug.

Any of the newly added tests for the Map strategy would have caught the
bug, as can be proven by removing the fix which makes the tests fail.