Add docker-compose file to start RAFT clusters.
This commit adds a new folder, docker-compose-raft/ which
includes a Jinja2 template and a bash script which allows
creating dynamic docker-compose files to start Weaviate
multinode clusters compatible with the RAFT protocol.

The only input parameter required is the number of voters,
by default there will be 2. The following command would
create a docker-compose-raft.yml file with 5 voters and
an extra Weaviate service which can be scaled:
$ docker-compose-raft/raft_cluster.sh 5