make object count in nodes API "async"
by "async" we mean that it only reflects objects that have already been
flushed into segments. It skips memtables. As a result, we don't ever
have to do complex calculations which makes this great for observability
purposes because it's super cheap to retrieve. The downside is that the
values take some time to reflect reality, in the worst case 60s after
the last write until the last memtable is flushed.

see also #4193