gh-4615 prevent deadlock when goroutine panics
The previous logic would assume that either routines can't exit and
there is always a routine to listen to the channel or that they exited
because a context was canceled. Since the errgroup was already using a
WithContext-wrapper, we could easily make sure that a panic cancels the
returned context. Now the logic works again and the select statement can
make progress – even when a routine in the error group panicked.

fixes #4615