ISSUE-4234 Add pagination to Get Tenants endpoint
Updated GetConsistentTenants endpoint with `after` and `limit` vars in
order to implement pagination for th endpoint. This required
refactoring the following: the implementation of QueryTenantRequest by
adding the `after` and `limit` vars. Also, the implementation of
getTenants (original handler.getTenant and cluster/store schema
.getTenants) method with the addition of `after` and `limit` vars  which
included the addition of implementation of pagination in its code
definition. Note, we now create the tenant slice based on the limit
size, or if it's not set, the len of `ss.Physical`.

`after` is the tenant for which GET of tenants will be AFTER.
`limit` is the limit of tenants we GET.

Added test for GetConsistenTenants endpoint.

Added additional test case for QueryTenants test case in cluster store
test for TestServiceEndpoints to ensure after and limit worked correctly
with the cluster store as well.