add support for sending azure deployment id and resource name via request headers
What's being changed:

When using Azure OpenAI, the deployment had to be configured in the vectorizer config when setting up the schema.
With this change, it is possible to omit it in the schema, or overwrite the setting on a per request base, similar to how you can set a different Azure API Key, via `X-Azure-Deployment-Id` and `X-Azure-Resource-Name` headers.

In addition to this, there is a new config option for openai modules `isAzure` which specifies whether the vectorizer should use Azure, and prevents developers from having to setup a spoof resourceName or deploymentId when setting up the schema. It should also lead to issues such as this one #4649 becoming less likely, as errors should then be about Azure OpenAI and not OpenAI.

This change implements #3605