Delay compaction in favor of not blocking user requests
The flush process itself does not block the user, however it holds a
maintenanceLock.RLock() which blocks compactions. That alone is fine,
however, because of the write-preferring nature of RWMutexes in Go, the
moment a compaction tries to obtain a Lock it will start blocking every
new request as long as there are old requests going on.

This commit adds a creative (read: hacky, but safe) solution to address
this:

* The compaction requires a `Lock()` on the RWMutex
* Go’s RWMutex is write-preferring, so the moment the compaction asks
  for this locks all future user requests are blocked
* However, the compaction is an “optional” background process, and it
  should not block the users
* So I added a (lock-free) “wait for flushing to complete” delay into
  the compaction that will not even try to obtain the lock if it knows
  that a flushing process is ongoing
* This definitely feels like a workaround to something that should be
  designed differently in the first place, but the only-side effect it
  has, is that the compaction is waiting a bit, so it adds no risk of
  new races, etc.

In addition, this commit adds a few warnings logs if specific logs are
held for too long.