Paged cache locking (#6358)
The vector cache lock is modified to reduce the number of locks/unlocks needed, especially for the flat index search.

The assignment of locks to IDs is changed:

Currently, every node is assigned to lock nodeId % 512 now, every node is assigned to (nodeId / pageSize) % 512.
This means that each 512 adjacent nodes share the same lock, reducing the number of locks/unlocks when iterating over all nodes. For HNSW pageSize is set to 1 so the locking strategy is the same as before.

Additionally a `GetAllInCurrentLock` function is added to to optimize iteration over the entire cache. It will skip cache
misses if the cache size is 1e12 (the default) as cache misses would be unexpected in this scenario.

---------

Co-authored-by: John Trengrove <john@weaviate.io>