Add new workflow to build ARM image with a comment.
One limitation with the current approach is that, to save
time, we are not building ARM images. However, for testing purposes
this adds some extra steps if we want to test an image from
a colleague on MacOS. Having to clone the branch into antoher
starting with build* and wait for all the CI jobs to finish.

This new workflow would get trigger upon the comment /build_arm
and if all the workflows have passed, then it will trigger the
Push-Docker job. Leaving the decission of building it at will.