Add new option --arm64-only into push-docker script.
This commit adds a new option to build docker images and push
then to Dockerhub.
Also, changes the behavior when a pull-request is created to
leverage the matrix-strategy to run both, amd and arm build jobs.
This way, if somebody needs an Intel based image the job will take
less time to build and won't be blocked if the ARM build fails.
ARM images will be build in ARM runners, removing the need for
virtualization and to keep one tag for both images, we push each
tag to an tag-arch (for example, 1.30.0-dev-92b06bf-arm64 and
1.30.0-dev-92b06bf-amd64) tag.

Also, it fixes the segmentation fault caused during multi-arch
docker builds. The cause was a cross-platform issue when building
musl on Alpine, requiring CGO_ENABLED=0:
https://docs.docker.com/docker-hub/image-library/trusted-content/#alpine-images