avoid nil logger in pv-pair
The current logic was far to complex and tried to pass a logger to pv.
That failed when a new, nested pv was created that could have a nil
logger.

This new solution is much simpler because it removes the logger from pv
entirely. Instead it takes the logger out of the Searcher struct which
is a singleton and always has a logger