Rotational quantization.
This patch adds support for the extended RaBitQ algorithm that quantizes
a D-dimensional vector of float32's to a D-dimensional vector of B-bit
integers. The quantization algorithm is based on encoding the original
vector to the nearest point in a set of randomly rotated points from an
evenly spaced D-dimensional grid that have been scaled to unit length.

An unoptimized gonum implementation of random rotation matrices is
included as a stand-alone file. Tests verify the properties of random
rotations.

A benchmark to test the recall of different quantization techniques is
also included. It is based on local HDF5 files in the ann-benchmarks
format. These have to be supplied to run the benchmark. Remember to
configure the go compiler to include the hdf5 headers. In the folder
compressionhelpers/figures there is a Python notebook that parses the
json output from go bench in order to create ASCII tables of the
benchmark results.

The implementation of RaBitQ comes with an option to pass a list of
centers. If given a list of centers the encoding algorithm will center
and normalize the point to be encoded around its nearest center.

This is a prototype-quality implementation. It needs to be optimized
for performance and precision, and the quality of the estimators should
be double-checked.