Replaced RaBitQ with fast rotational quantization.
Removed the custom encoding struct and switched to using a byte slice.
Internally the rotational quantization converts back and forth between
the byte slice and an RQCode struct. This could probably be made cleaner
and more performant.

A minimal set of unit tests are in place for rotational quantization,
but more should be added to cover corner cases and build confidence in
the encoding.