harden tests and change test plan depending on race detector
The full test with all edge cases including on larger datasets is not
feasible with the race detector on, but it's very valuable and has
identified many bugs.

Therefore we want both. When the race detector is on, run a minimal test
suite. When it's off, run the full test suite. We do both on CI.