chore(runtime): Improve log message and fix flaky test
This fixes two problems:
1. replace `%w` -> `%v` for better printing of errors.
2. Fix flaky tests with config reload.

(2) was tricky to find. Turns out the problem is with filesystem cache. The race between file writer and config manager (reader).
Somtimes the reader reads the file as empty due to this race. I verified this with printing the hash changes. The fix is to use `tmpfile + rename()` to do
it atomically at filesystem level.

Thanks Etienne for pointing it out.

Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>