Benchmarked and optimized the fast rotation.
The fast rotation now supports an in-place rotation that is intended to
be used with a pool of float64 vectors when encoding data vectors.

As part of this patch I investigated whether the rotation can be
performed directly using float32 values, but the floating point error is
roughly an order of magnitude greater compared to using float64.

Acquiring a float64 vector from a pool should be neglible compared to
the time it takes to perform the rotation itself.

I also investigated the number of rounds for the rotation, and it does
seem like we need at least 5 rounds for short vectors before the output
distribution closely resembles the normal distribution.

I removed the restriction to only transform blocks of size 2^2k since
the effects on the error seemed minimal compared to using blocks of size
2^k.