Simplifying the block-processing approach.
Modified the patch to use a slightly more minimal loop and fewer bit
tricks while keeping the performance improvement.

The original patch gives as much as a 3x improvement in the encoding
time on vectors of length 4096! It seems like much of the performance
improvement is enabled by two modifications: 1) avoiding direct writes
into code slice for every entry in the vector to be encoded, and 2)
shifting the indicator bit one by one when processing each position in
the block rather than using something like 1 << j explicitly.