Rotational Quantization (#8260)
* Rotational quantization.

This patch adds support for the extended RaBitQ algorithm that quantizes
a D-dimensional vector of float32's to a D-dimensional vector of B-bit
integers. The quantization algorithm is based on encoding the original
vector to the nearest point in a set of randomly rotated points from an
evenly spaced D-dimensional grid that have been scaled to unit length.

An unoptimized gonum implementation of random rotation matrices is
included as a stand-alone file. Tests verify the properties of random
rotations.

A benchmark to test the recall of different quantization techniques is
also included. It is based on local HDF5 files in the ann-benchmarks
format. These have to be supplied to run the benchmark. Remember to
configure the go compiler to include the hdf5 headers. In the folder
compressionhelpers/figures there is a Python notebook that parses the
json output from go bench in order to create ASCII tables of the
benchmark results.

The implementation of RaBitQ comes with an option to pass a list of
centers. If given a list of centers the encoding algorithm will center
and normalize the point to be encoded around its nearest center.

This is a prototype-quality implementation. It needs to be optimized
for performance and precision, and the quality of the estimators should
be double-checked.

* add integration of rq into weaviate

* add comments to not implemented functions

* add new parameters to rotational quantization

* add rq configuration in the schema

* Replaced RaBitQ with fast rotational quantization.

Removed the custom encoding struct and switched to using a byte slice.
Internally the rotational quantization converts back and forth between
the byte slice and an RQCode struct. This could probably be made cleaner
and more performant.

A minimal set of unit tests are in place for rotational quantization,
but more should be added to cover corner cases and build confidence in
the encoding.

* add additional function for integration

* run swagger

* run linter

* minor change

* Fixed failing RQ unit test.
Added estimation error concentration bound test.

* Quick and dirty thread-safety fix!

Also added the flat index recall benchmarks again.

* add inferred dimension and enable the compression post creation

* remove old print for debugging

* add rescoring false by default for rq

* minor changes

* add rescoring limit

* Benchmarked and optimized the fast rotation.

The fast rotation now supports an in-place rotation that is intended to
be used with a pool of float64 vectors when encoding data vectors.

As part of this patch I investigated whether the rotation can be
performed directly using float32 values, but the floating point error is
roughly an order of magnitude greater compared to using float64.

Acquiring a float64 vector from a pool should be neglible compared to
the time it takes to perform the rotation itself.

I also investigated the number of rounds for the rotation, and it does
seem like we need at least 5 rounds for short vectors before the output
distribution closely resembles the normal distribution.

I removed the restriction to only transform blocks of size 2^2k since
the effects on the error seemed minimal compared to using blocks of size
2^k.

* store uncompressed query in distancer

* add persistance for rotational quantization

* Optimized distance computations for RQ.

Optimized the RQ distance functions by working directly on the byte
array using small helper functions that appear to be inlined according
to local benchmarks.

Also removed the switch statement in distance computations and instead
used indicator variables to avoid any branching. This gave a further
speedup of ~1 nanosecond in benchmarks. The RQ distance computations
should now be slightly faster than SQ distance computations.

* add on persistance for rq

* run linter

* update user config tests

* update test default values

* update default values for rq

* minor changes in test

* fix test

* fix tests

* add rq to snapshotting

* add tests for condensor and deserializer

* Slight optimization of the fast rotation.

* Speeding up RQ encoding.

Reducing the number of rotation rounds from 5 to 3. Added a constant in
an attempt at tracking our loss of estimation accuracy from using
heuristics compared to extended RaBitQ

* add multivector support

* add locking mechansim for multivec + rq

* changes in persistance

* minor changes

* update rescoring logic

* Added additional benchmark experiments.

Investigation of the effects of variations on plain RQ.

* Centered rotational quantization prototype.

Protyping and experimenting with truncated and centered RQ.

* Truncated rotational quantization.

Added prototype implementation of a variation of RQ that stores the
largest entries of the rotated vector seperately. It improves recall
slightly, even in the >99% recall regime, but it comes at a cost of
increase space usage and distance computation time.

Overall the benefits seem smaller than centering.

* restore distance parameters for rotational quantization

* add sync once when restore RQ compressor

* run swagger

* remove data bits and query bits

* add new compressed quantizer distancer

* add rotational quantizer flag

* Optimized fast rotation.

Added various optimizations to the fast rotations without hurting
quality too much.

Some of the changes include:
- Baking normalization into the transform.
- Specialized implementation of length 64 and length 256 transforms.
- Restricting the max length of the block transforms to 256.
- Switching to float32 instead of float64 (did not improve performance
  on my machine, but should generally be more efficient and amenable to
  later SIMD optimization, as well as using less memory).
- Performing the random swaps with greater locality.
- Storing the random signs directly as float32 values to avoid casts.

Overall this reduces encoding time by around 40%.

* restore compressor for multivector

* update deserializer for signs

* add muvera state on startup

* run linter on hdf5_test.go

* Cleanup before integration.

Removing experimental features and benchmarks.

Adding some simplifications and code reuse.

Gained a further ~5% speedup of the encoding from simplifying the
rounding and using our fast dot product implementation to compute the
vector norm.

* update snapshot metadata for rq

* Code point distribution test.

Added a test to verify that our encoding uses the full byte.
No changes to the algorithm.

* Fix RQ acceptance test

* update bytes read on deserialization

* fix test

* Cleanup of fast rotation.

Minimizing the exported functionality to what we actually use.

The specialized Fast Walsh Hadamard Transforms are exported for testing
only using the export_test.go construct.

Removed the fully random rotation functionality which unused.

Updated fast rotation benchmark that measures how closely the output
of the fast rotation resembles a randomly rotated vector.

* Explicit and graceful handling of abnormal input.

Abnormal input to the encode function such as the nil vector will now
always explicitly return the code for the zero vector. This is not
supposed to happen under normal circumstances. Vectors that are too long
will be truncated to the output dimension supported by the quantizer.

* use default seed for fast rotation

* tidy go mod file

* minor fix on test

* validate bits for rq

* update fixed seed

* format update config

---------

Co-authored-by: Tobias Christiani <tobias@weaviate.io>
Co-authored-by: John Trengrove <john@weaviate.io>