refactor(repos/db): start nodeWideMetricsObserver whenever Prometheus is configured
This module is responsible for collecting and publishing all
group-able shard-level metrics. That way we centralize control
over if and how these metrics are aggregated in one place
rather than forcing each shard to always check "if prom.Group".

The only time a shard is required to update its metrics is when
it is being shut down or dropped, as nodeWideMetricsObserver
has no way of learning that at the moment.