chore(alias): Objects api response will have original class name
Currently, when objects were create/updated/delete/listed via alias name, that will contain `class` key with value as alias name
but that is confusing and we decided not a best approach.
It's better to have original class name as value for `class` key no matter how we access the objects (source of truth).

This PR make that class key consistent in all those objects endpoints

Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>