Added a compressed distancer.
We have a special interface for creating a distancer when deleting nodes
in the HNSW graph. We were restoring a float32 representation from the
1-bit quantization and creating a regular distancer. This may have
resulted in some bias when performing distance estimation in connection
with repairing a graph after deletions.

In order to produce better distance estimates we instead return a custom
distancer for compressed vectors that relies on
DistanceBetweenCompressedVectors() which we have unit tests for.