raft: Fix skipping the error return from leader
If query forwarded to leader from the follower failes in leader, those
errors are skipped to be propaged to the follower.

Identified this issue when testing alias with RBAC + OIDC. Thanks to
@Jose :)

Basically, on the follower side, we should stop as soon as alias is not
found. but Follower ignores this error (NOT FOUND) from the leader

Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>