make locking in bucket.Put more fine-grained
Similar to bucket.Get, we don't need to hold the flushLock during the
entire duration of the write. What we do need is the guarnatee that the
current memtable will not be flushed before the write has finished
because otherwise we could lose the write.

Previously the flushing memtable was immutable. This commit changes this
behavior. It stays mutable, but only for writers that started writing
while the memtable was still active. This way we can guarantee that the
reference count will eventually drop zero because the active->flushing
switch means new writers will reference the (new) active memtable.

The mutability is not an issue because the flush only happens after the
writer count has reached zero. At that point the memtable is immutable.
So the flushing logic itself does not change.