[target: copy on read] extend LSM store with summarized slow log (#9353)
This extends the existing slow log with some LSM internal stats for the GetBySecondary call. This endpoint is called once per object in a search both for rescore and retrieval. So if you have a limit of 100 and use compression, this endpoint would be called 200 times. To prevent the slow logs from exploding in size, the stats are reduced to some percentiles, min, and max.

Original commits:
* include lsm store internals in slow log
* remove debug log
* print bucket slow log stats instead of individual values
* fix linter
* add license headers