Support for reading float32 vectors from parquet. (#9388)
This patch modifies how the training data vectors are read. Instead of
casting byte arrays we read the parquet values stored in the list
representation from the parquet column and use the repetition level to
build our collection of vectors. Currently we only support float32
vectors but this approach can be extended to other Parquet by reading
their value from the parquet.

It seems that this is slightly less performant than casting byte arrays.
This patch includes a benchmark that shows that reading ~54K 768-dim
vectors takes approximately 275ms. The benefits of this format is that
it makes the datasets integrate well with the huggingface ecosystem.
I.e. datasets can be downloaded, processed, uploaded, and viewed online
without any special knowledge about the encoding.

In the (relatively small) benchmark we use ~6.6 nanoseconds per float32
value when reading training data. If we interpolate to 1536-dimensional
vectors we can read approximately 100K vectors per second. This is at
least an order of magnitude faster than our indexing time, so this
overhead should be tolerable.

In this iteration we load the train dataset by searching for a file
named train.parquet in the huggingface repo and downloading it, and
similarly for the test datast. This currently means we have one dataset
per repo, and one file per configuration/split. We may need to update
this to e.g. support sharding or multiple splits, but it is easier to
see the best structure once we have produced more datasets.