Added a streaming/chunk reader for parquet data.
Attempted to refactor to increase code re-use and add more validation.
Switched back to using byte arrays for the vectors/embeddings. This
gives a performance of ~1ns per float32 value (on an M4 Pro) or only
~160ms for reading the 100k OpenAI vectors.

The parquet file format produced by the huggingface library is a bit
messy in the sense that all fields are optional. We may have to switch
over to using pyarrow when saving our datasets in order to get more
precise control so that we do not have to validate as much.

If we could get a performant solution with REQUIRED id and vector
columns and the FIXED_LEN_BYTE_ARRAY type for the vector that we could
read using parquet-go's generic reader so that we would also get schema
validation, that would be nice.

Unfortunately it seems like we still have to make the neighbors column
optional if we want to use the same huggingface configuration for test
and train, but we need to experiment further with the exact limitations
here.