Added DataReader support for parquet row slices.
The benchmarker supports reading different sections of the rows in a
dataset. This is used when loading the first rows of a dataset for
training a quantizer and later reading all of the remaining rows. In
order to provide the same functionality for parquet files I added
support for reading only the rows in the interval startRow:endRow that
is specified when constructing the DataReader.

Removed logging from the dataset because logs were emitted every time
we opened a file, which we do multiple times during benchmarking.

Moved the DataReader into its own file and added small local parquet
files for basic unit tests.