Initial support for Huggingface ANN datasets (#9298)
* Add new Huggingface parquet dataset

* Include Test and Train splits

* Put tests under benchmark tag

* Add column validation

* Move allocation out of loop

* Fix linting

* Support for reading float32 vectors from parquet. (#9388)

This patch modifies how the training data vectors are read. Instead of
casting byte arrays we read the parquet values stored in the list
representation from the parquet column and use the repetition level to
build our collection of vectors. Currently we only support float32
vectors but this approach can be extended to other Parquet by reading
their value from the parquet.

It seems that this is slightly less performant than casting byte arrays.
This patch includes a benchmark that shows that reading ~54K 768-dim
vectors takes approximately 275ms. The benefits of this format is that
it makes the datasets integrate well with the huggingface ecosystem.
I.e. datasets can be downloaded, processed, uploaded, and viewed online
without any special knowledge about the encoding.

In the (relatively small) benchmark we use ~6.6 nanoseconds per float32
value when reading training data. If we interpolate to 1536-dimensional
vectors we can read approximately 100K vectors per second. This is at
least an order of magnitude faster than our indexing time, so this
overhead should be tolerable.

In this iteration we load the train dataset by searching for a file
named train.parquet in the huggingface repo and downloading it, and
similarly for the test datast. This currently means we have one dataset
per repo, and one file per configuration/split. We may need to update
this to e.g. support sharding or multiple splits, but it is easier to
see the best structure once we have produced more datasets.

* Add back subset split

* Added a streaming/chunk reader for parquet data.

Attempted to refactor to increase code re-use and add more validation.
Switched back to using byte arrays for the vectors/embeddings. This
gives a performance of ~1ns per float32 value (on an M4 Pro) or only
~160ms for reading the 100k OpenAI vectors.

The parquet file format produced by the huggingface library is a bit
messy in the sense that all fields are optional. We may have to switch
over to using pyarrow when saving our datasets in order to get more
precise control so that we do not have to validate as much.

If we could get a performant solution with REQUIRED id and vector
columns and the FIXED_LEN_BYTE_ARRAY type for the vector that we could
read using parquet-go's generic reader so that we would also get schema
validation, that would be nice.

Unfortunately it seems like we still have to make the neighbors column
optional if we want to use the same huggingface configuration for test
and train, but we need to experiment further with the exact limitations
here.

* Changed name of embedding column to vector.

Minor updates from go mod tidy.

* Added DataReader support for parquet row slices.

The benchmarker supports reading different sections of the rows in a
dataset. This is used when loading the first rows of a dataset for
training a quantizer and later reading all of the remaining rows. In
order to provide the same functionality for parquet files I added
support for reading only the rows in the interval startRow:endRow that
is specified when constructing the DataReader.

Removed logging from the dataset because logs were emitted every time
we opened a file, which we do multiple times during benchmarking.

Moved the DataReader into its own file and added small local parquet
files for basic unit tests.

* Fix headers

* Fix swagger

* Add new weaviate HF repo

* Remove neighbor length assumption

---------

Co-authored-by: Tobias Christiani <tobias@weaviate.io>