Mitigation/logging for PQ out of bounds panic (issue #9470)
Testing has revealed an issue (#9470) where the k-means algorithm used
by product quantization is passed either nil vectors or empty vectors.
This is a difficult issue to reproduce and it may stem from an
underlying race condition.

This patch adds additional verification that all vectors passed to the
PQ training algorithm have the same expected length. Prior to this patch
we only checked if vectors returned by the sharded lock cache were nil.

This patch also adds locking around the return of the cache's internal
slice of vectors in the implementation of All(). This method is only
called in the context of compression. Prior to this patch there was no
synchronization/protection when calling All() except for the HNSW
compressActionLock. However it appears that AddBatch() can trigger a
growth of the cache which will swap out the underlying cache slice
without acquiring the compressActionLock. It is unlikely that this is
the cause of the error we are seeing, but not impossible.

In the logs for the issue we see that the issue happens when the k-means
algorithm is passed a slice with the following signature:

initializeRandom(0x40005c83c0, {0x4012f00000, 0x2710, 0x3ebe}

The slice has a length of 10000 (0x2710) and a capacity of 16062
(0x3ebe). Since the slice is initially instantiated to have the same
capacity as the length of the vector cache it is unclear how it can have
grown to exceed this and this might indicate a race condition or a bug
in the sampling algorithm.

I cannot identify any other serious concurrency issues in the sharded
lock cache itself by inspecting the code, so I suspect that the problem
may be that the underlying VectorForId function that handles cache
misses can sometimes return a vector that does not have the right length or
that it is not thread-safe.

It is also possible that we have overlooked a bug in the k-means
implementation, but then it should have been possible to reproduce the
issue in a unit test and we would likely have seen the issue more often.